Time: 32 ms (13.81%), Space: 9 MB (67.04%) - LeetHub

+++ 1442-count-triplets-that-can-form-two-arrays-of-equal-xor/1442-count-triplets-that-can-form-two-arrays-of-equal-xor.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int countTriplets(vector<int>& arr) {
-        long a=0,b=0,count=0;
+        int a=0,b=0,count=0;
         for(int i=0;i<arr.size();i++)
         {  a=arr[i];
             for(int j=i+1;j<arr.size();j++)
